+++ game/world/src/loader/terrain_source/mod.rs
@@ -112,7 +112,7 @@ impl TerrainSource {
         &self,
         chunks: impl Iterator<Item = ChunkLocation>,
         z_range: (GlobalSliceIndex, GlobalSliceIndex),
-        per_point: impl FnMut(u32, WorldPosition),
+        per_point: impl FnMut(u64, WorldPosition),
     ) {
         match self {
             TerrainSource::Memory(_) => {}
